WordFactory, ByteWordFactory & LongWordFactory should work as intended. Might need a redo further down the line when we figure the rest out.

+++ src/computer/LongWord.java
@@ -6,7 +6,7 @@ public class LongWord extends Word{
 	
 	private long value;
 	
-	public LongWord(int length) {
+	public LongWord(long length) {
 	}
 
 	@Override
